+++ src/spyglass/overlay/pane/status_bar.h
@@ -4,7 +4,8 @@ namespace spyglass {
 
 class Capture;
 struct BytesView;
+struct PacketList;
 
-void draw_status_bar(const Capture &capture, const BytesView &bytes);
+void draw_status_bar(const Capture &capture, const PacketList &list, const BytesView &bytes);
 
 }  // namespace spyglass