+++ src/spyglass/overlay/pane/toolbar.h
@@ -3,7 +3,8 @@
 namespace spyglass {
 
 class Capture;
+struct Filter;
 
-void draw_toolbar(Capture &capture);
+void draw_toolbar(Capture &capture, const Filter &filter, bool &filter_open);
 
 }  // namespace spyglass